move the writing system id tests into their own test class
